+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed closing of database/sql connection (aka `YDB` session)
+
 ## v3.42.4
 * Added `ydb.WithDisableServerBalancer()` database/sql connector option